+++ tests/static_tests/performance/test_overall_performance.py
@@ -83,7 +83,15 @@ def test_performance_element_initialisation(mocked_selenium_driver, case, set_el
     if sys.version_info >= (3, 11):
         expected_peak_mem = 4.8
     if sys.version_info >= (3, 12):
+        expected_peak_mem = 3.8
+        expected_init_duration = 0.4
+    if sys.version_info >= (3, 13):
+        expected_peak_mem = 4.0
+        expected_init_duration = 0.4
+    if sys.version_info >= (3, 14):
         expected_peak_mem = 5.0
+        expected_init_duration = 0.4
+        init_without_profiling_expected = 0.18
 
     assert init_without_profiling_stop_timestamp < init_without_profiling_expected,\
         f'Execution without profiling takes too much time: {init_without_profiling_stop_timestamp}'